+++ includes/globals.js
@@ -3,6 +3,7 @@ var last_selected_element = null;
 var serverTimeOffset = 0;
 var resizeTimer = null;
 var menuOpeners = new Array();
+var menuClosers = new Array();
 const masonry_gutter = 8;
 var coverscraper;
 var lastfm;