+++ includes/vars.php
@@ -173,9 +173,6 @@ $prefs = array(
     "hidebrowser" => false,
     "shownupdatewindow" => '',
     "scrolltocurrent" => false,
-    "alarmtime" => 43200,
-    "alarmon" => false,
-    "alarmramp" => false,
     "alarm_ramptime" => 30,
     "alarm_snoozetime" => 8,
     "lastfmlang" => "default",
@@ -230,16 +227,7 @@ $prefs = array(
     "podcast_sort_2" => 'Category',
     "podcast_sort_3" => 'new',
     "bgimgparms" => (object) array('dummy' => 'baby'),
-    "alarmrepeat" => false,
-    "alarmrepeat_monday" => false,
-    "alarmrepeat_tuesday" => false,
-    "alarmrepeat_wednesday" => false,
-    "alarmrepeat_thursday" => false,
-    "alarmrepeat_friday" => false,
-    "alarmrepeat_saturday" => false,
-    "alarmrepeat_sunday" => false,
-    "alarmplayitem" => false,
-    "alarm_itemtoplay" => ''
+    "alarms" => array( )
 );
 
 // Prefs that should not be exposed to the browser for security reasons