+++ docs/Changelog.md
@@ -9,6 +9,7 @@ Note that some versions listed here may be unreleased, I use version number incr
 * [Spoken Word](/RompR/Spoken-Word) support - for things like Audiobooks where you don't want them to be in your 'Music' Collection.
 * A few small bugfixes
 * Loading of Custom Background Images is now much faster
+* Big changes to the [Alarm Clock](/RompR/Alarm-And-Sleep). You can now have multiple alarms, they can be set to repeat on specific days of the week, and can play a specific item - an album, radio station, playlist, or whatever you like. If you've been using the alarm clock, you should read the docs as everything has changed.
 
 ## Version 1.25